refactor: isolate Pi reasoning patch
